@@ -1,10 +1,11 @@
 library one_context;
 
-export 'package:one_context/src/controllers/one_context.dart' show OneContext;
+export 'package:one_context/src/components/one_context_widget.dart';
 export 'package:one_context/src/components/one_notification.dart'
     show OneNotification, OnVisited, NotificationBuilder;
 export 'package:one_context/src/components/one_notifier.dart'
     show OneNotifier, NotificationPayload;
+export 'package:one_context/src/controllers/one_context.dart' show OneContext;
+export 'package:one_context/src/controllers/one_platform.dart' show OnePlatform;
 export 'package:one_context/src/controllers/one_theme_controller.dart'
     show OneThemeController, OneThemeChangerEvent;
-export 'package:one_context/src/controllers/one_platform.dart' show OnePlatform;
