@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' as mat;
 import 'package:one_context/src/controllers/one_context.dart';
 
 typedef Widget DialogBuilder(BuildContext context);
@@ -9,6 +10,16 @@ mixin DialogController {
 
   /// The current context
   BuildContext? get context => OneContext().context;
+
+  GlobalKey<ScaffoldState>? _scaffoldKey;
+  GlobalKey<ScaffoldState> get scaffoldKey {
+    _scaffoldKey ??= GlobalKey<ScaffoldState>();
+    return _scaffoldKey!;
+  }
+
+  set scaffoldKey(scaffKey) => _scaffoldKey = scaffKey;
+  BuildContext? get _scaffoldContext => _scaffoldKey?.currentContext;
+  ScaffoldState? get _scaffoldState => _scaffoldKey?.currentState;
 
   ValueNotifier<List<Widget>> _dialogs = ValueNotifier([]);
   ValueNotifier<List<Widget>> get dialogNotifier => _dialogs;
@@ -36,48 +47,6 @@ mixin DialogController {
     _dialogs.value.clear();
   }
 
-  Future<T?> Function<T>({
-    required Widget Function(BuildContext) builder,
-    bool? barrierDismissible,
-    bool useRootNavigator,
-    Color? barrierColor,
-    String? barrierLabel,
-    bool useSafeArea,
-    RouteSettings? routeSettings,
-    Offset? anchorPoint,
-  })? _showDialog;
-
-  Future<T?> Function<T>({
-    required Widget Function(BuildContext) builder,
-    Color? backgroundColor,
-    double? elevation,
-    ShapeBorder? shape,
-    Clip? clipBehavior,
-    bool isScrollControlled,
-    bool useRootNavigator,
-    bool isDismissible,
-    BoxConstraints? constraints,
-    Color? barrierColor,
-    bool? enableDrag,
-    RouteSettings? routeSettings,
-    AnimationController? transitionAnimationController,
-    Offset? anchorPoint,
-  })? _showModalBottomSheet;
-
-  ScaffoldFeatureController<SnackBar, SnackBarClosedReason> Function(
-      SnackBar Function(BuildContext?) builder)? _showSnackBar;
-
-  PersistentBottomSheetController<T> Function<T>({
-    Widget Function(BuildContext) builder,
-    Color? backgroundColor,
-    double? elevation,
-    ShapeBorder? shape,
-    Clip? clipBehavior,
-    BoxConstraints? constraints,
-    bool? enableDrag,
-    AnimationController? transitionAnimationController,
-  })? _showBottomSheet;
-
   /// Displays a Material dialog above the current contents of the app, with
   /// Material entrance and exit animations, modal barrier color, and modal
   /// barrier behavior (dialog is dismissible with a tap on the barrier).
@@ -87,27 +56,28 @@ mixin DialogController {
     String? barrierLabel,
     RouteSettings? routeSettings,
     Offset? anchorPoint,
-    bool barrierDismissible = true,
+    bool? barrierDismissible = true,
     Color? barrierColor = Colors.black54,
     bool useSafeArea = true,
   }) async {
     if (!(await _contextLoaded())) return null;
 
     Widget dialog = builder(context!);
-    if (barrierDismissible == true) addDialogVisible(dialog);
+    addDialogVisible(dialog);
 
-    return _showDialog!<T>(
-      builder: (_) => dialog,
-      barrierDismissible: barrierDismissible,
-      useRootNavigator: useRootNavigator,
-      barrierColor: barrierColor,
-      barrierLabel: barrierLabel,
-      useSafeArea: useSafeArea,
-      routeSettings: routeSettings,
-      anchorPoint: anchorPoint,
-    ).whenComplete(() {
-      if (barrierDismissible == true) removeDialogVisible(widget: dialog);
-    });
+    return mat
+        .showDialog<T>(
+          context: _scaffoldContext!,
+          builder: (_) => dialog,
+          barrierDismissible: barrierDismissible ?? true,
+          useRootNavigator: useRootNavigator,
+          barrierColor: barrierColor,
+          barrierLabel: barrierLabel,
+          useSafeArea: useSafeArea,
+          routeSettings: routeSettings,
+          anchorPoint: anchorPoint,
+        )
+        .whenComplete(() => removeDialogVisible(widget: dialog));
   }
 
   /// Removes the current [SnackBar] by running its normal exit animation.
@@ -133,7 +103,8 @@ mixin DialogController {
   Future<ScaffoldFeatureController<SnackBar, SnackBarClosedReason>?>
       showSnackBar({required SnackBar Function(BuildContext?) builder}) async {
     if (!(await _contextLoaded())) return null;
-    return _showSnackBar!(builder);
+    return ScaffoldMessenger.of(_scaffoldContext!)
+        .showSnackBar(builder(context));
   }
 
   /// Shows a modal material design bottom sheet.
@@ -159,33 +130,30 @@ mixin DialogController {
     if (!(await _contextLoaded())) return null;
 
     Widget dialog = builder(context!);
-    if (isDismissible == true) addDialogVisible(dialog);
+    addDialogVisible(dialog);
 
-    return _showModalBottomSheet!<T>(
-      builder: builder,
-      backgroundColor: backgroundColor,
-      clipBehavior: clipBehavior,
-      elevation: elevation,
-      isDismissible: isDismissible,
-      isScrollControlled: isScrollControlled,
-      shape: shape,
-      useRootNavigator: useRootNavigator,
-      constraints: constraints,
-      barrierColor: barrierColor,
-      enableDrag: enableDrag,
-      routeSettings: routeSettings,
-      transitionAnimationController: transitionAnimationController,
-      anchorPoint: anchorPoint,
-    ).whenComplete(() {
-      if (isDismissible == true) removeDialogVisible(widget: dialog);
-    });
+    return mat
+        .showModalBottomSheet<T>(
+          context: _scaffoldContext!,
+          builder: builder,
+          backgroundColor: backgroundColor,
+          clipBehavior: clipBehavior,
+          elevation: elevation,
+          isDismissible: isDismissible,
+          isScrollControlled: isScrollControlled,
+          shape: shape,
+          useRootNavigator: useRootNavigator,
+          constraints: constraints,
+          barrierColor: barrierColor,
+          enableDrag: enableDrag = true,
+          routeSettings: routeSettings,
+          transitionAnimationController: transitionAnimationController,
+          anchorPoint: anchorPoint,
+        )
+        .whenComplete(() => removeDialogVisible(widget: dialog));
   }
 
-  /// Shows a material design bottom sheet in the nearest [Scaffold] ancestor. If
-  /// you wish to show a persistent bottom sheet, use [Scaffold.bottomSheet].
-  ///
-  /// Returns a controller that can be used to close and otherwise manipulate the
-  /// bottom sheet.
+  /// Shows a persistent bottom sheet
   Future<PersistentBottomSheetController<T>?> showBottomSheet<T>({
     required Widget Function(BuildContext) builder,
     Color? backgroundColor,
@@ -196,9 +164,10 @@ mixin DialogController {
     bool? enableDrag,
     AnimationController? transitionAnimationController,
   }) async {
-    if (!(await _contextLoaded())) return null;
-    return _showBottomSheet!<T>(
-      builder: builder,
+    if (!(await _scaffoldContextLoaded())) return null;
+
+    return _scaffoldState!.showBottomSheet(
+      builder,
       backgroundColor: backgroundColor,
       elevation: elevation,
       shape: shape,
@@ -209,60 +178,17 @@ mixin DialogController {
     );
   }
 
-  /// Register callbacks
-  void registerCallback(
-      {Future<T?> Function<T>({
-        required Widget Function(BuildContext) builder,
-        bool? barrierDismissible,
-        bool useRootNavigator,
-        Color? barrierColor,
-        String? barrierLabel,
-        bool useSafeArea,
-        RouteSettings? routeSettings,
-        Offset? anchorPoint,
-      })?
-          showDialog,
-      Future<T?> Function<T>({
-        required Widget Function(BuildContext) builder,
-        Color? backgroundColor,
-        double? elevation,
-        ShapeBorder? shape,
-        Clip? clipBehavior,
-        bool? isScrollControlled,
-        bool? useRootNavigator,
-        bool? isDismissible,
-        BoxConstraints? constraints,
-        Color? barrierColor,
-        bool? enableDrag,
-        RouteSettings? routeSettings,
-        AnimationController? transitionAnimationController,
-        Offset? anchorPoint,
-      })?
-          showModalBottomSheet,
-      ScaffoldFeatureController<SnackBar, SnackBarClosedReason> Function(
-              SnackBar Function(BuildContext?) builder)?
-          showSnackBar,
-      PersistentBottomSheetController<T> Function<T>({
-        Widget Function(BuildContext)? builder,
-        Color? backgroundColor,
-        double? elevation,
-        ShapeBorder? shape,
-        Clip? clipBehavior,
-        BoxConstraints? constraints,
-        bool? enableDrag,
-        AnimationController? transitionAnimationController,
-      })?
-          showBottomSheet}) {
-    _showDialog = showDialog;
-    _showSnackBar = showSnackBar;
-    _showModalBottomSheet = showModalBottomSheet;
-    _showBottomSheet = showBottomSheet;
-  }
-
   /// Pop the top-most dialog off the OneContext.dialog.
   popDialog<T extends Object>([T? result]) async {
     if (!(await _contextLoaded())) return;
-    return Navigator.of(context!).pop<T>(result);
+    if (OneContext().hasDialogVisible)
+      return Navigator.of(_scaffoldContext!).pop<T>(result);
+  }
+
+  /// Pop the PersistentBottomSheet
+  popBottomSheet<T extends Object>([T? result]) async {
+    if (!(await _scaffoldContextLoaded())) return;
+    return Navigator.of(_scaffoldContext!).pop<T>(result);
   }
 
   Future<bool> _contextLoaded() async {
@@ -271,5 +197,14 @@ mixin DialogController {
       throw NO_CONTEXT_ERROR;
     }
     return OneContext.hasContext;
+  }
+
+  Future<bool> _scaffoldContextLoaded() async {
+    await Future.delayed(Duration.zero);
+    final contextIsNull = _scaffoldContext == null;
+    if (contextIsNull && !kReleaseMode) {
+      throw NO_CONTEXT_ERROR;
+    }
+    return !contextIsNull;
   }
 }
